Final Version of My College Project
Exercise Statement (The code was written in Portuguese because it's my college language):
Exercise 18

Implement the following code. Each class should be in a separate module (file):

Create a Person Class, using encapsulated attributes with their respective getters and setters, and also the default initializer. Attributes: name, address, and phone.

Implement the class method (create_anonymous) for the Person class that generates an instance with name="John Doe", address="Unknown", phone="Unknown".

Consider, as a subclass of the Person class (developed in the previous item), the Supplier class.

Consider that each instance of the Supplier class has, in addition to the attributes that characterize the Person class, the attributes:
credit_value – corresponding to the maximum credit granted/admitted by the supplier, and
debt_value – amount owed to the supplier.

Implement in the Supplier class, in addition to the getters and setters, a method:
get_balance that returns the difference between the values of the credit_value and debt_value attributes.

After implementing the Supplier class, create an appropriate test program that allows you to verify the functionality of the methods implemented in the Supplier class and those inherited from the Person class.

Consider, as a subclass of the Person class, the Employee class.

Consider that each instance of the Employee class has, in addition to the attributes that characterize the Person class, the attributes:
sector_code,
base_salary – base salary, and
tax – percentage of taxes withheld.

Implement the Employee class with getters and setters, and a calculate_salary method that returns the net salary of the employee.

Write a suitable test program for the Employee class.

Implement the Administrator class as a subclass of the Employee class.

A specific administrator has as attributes (in addition to the attributes of the Person class and the Employee class) the attribute:
travel_allowance – allowances for travel, stays, etc.

Note that you should redefine the inherited calculate_salary method in the Administrator class (the salary of an administrator is equivalent to the normal salary of an employee plus travel allowances).

Write a suitable test program for this class.

Implement the Worker class as a subclass of the Employee class.

A specific worker has as attributes, in addition to the attributes of the Person class and the Employee class, the attributes:
production_value – which corresponds to the monetary value of the items actually produced by the worker, and
commission – which corresponds to the percentage of the production_value that will be added to the worker's base salary.

Note that you should redefine the inherited calculate_salary method in this subclass (the salary of a worker is equivalent to the normal salary of an employee plus the commission).

Write a suitable test program for this class.

Implement the Salesman class as a subclass of the Employee class.

A specific salesman has as attributes, in addition to the attributes of the Person class and the Employee class, the attributes:
sales_value – corresponding to the monetary value of the items sold, and
commission – percentage of the sales_value that will be added to the salesman's base salary.

Note that you should redefine the inherited calculate_salary method in this subclass (the salary of a salesman is equivalent to the normal salary of an employee plus the commission).

Write a suitable test program for this class.

Implement and override the __str__() method in each of the classes.

Test the implemented functionalities.